# Backend/api_documentation.yaml
openapi: 3.1.0
info:
  title: CivicResolve AI
  version: 0.1.0
paths:
...
  /api/complaint/anonymous:
    post:
      tags:
      - public
      summary: File Anonymous Complaint
      operationId: file_anonymous_complaint_api_complaint_anonymous_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_file_anonymous_complaint_api_complaint_anonymous_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnonymousComplaintResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/complaint/track/{token}/resolution:
    put:
      tags:
      - public
      summary: Citizen Respond Resolution
      operationId: citizen_respond_resolution_api_complaint_track__token__resolution_put
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CitizenResolutionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Citizen Respond Resolution Api Complaint Track  Token  Resolution
                  Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/commissioner/complaint/{complaint_id}:
    get:
      tags:
      - commissioner
      summary: Commissioner Complaint Detail
      operationId: commissioner_complaint_detail_api_commissioner_complaint__complaint_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommissionerComplaintDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AnonymousComplaintResponse:
      properties:
        message:
          type: string
          title: Message
        trackingToken:
          type: string
          title: Trackingtoken
        complaintId:
          type: integer
          title: Complaintid
      type: object
      required:
      - message
      - trackingToken
      - complaintId
      title: AnonymousComplaintResponse
...
    Body_file_anonymous_complaint_api_complaint_anonymous_post:
      properties:
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        addressText:
          anyOf:
          - type: string
          - type: 'null'
          title: Addresstext
        categoryId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Categoryid
        photo:
          anyOf:
          - type: string
            contentMediaType: application/octet-stream
          - type: 'null'
          title: Photo
      type: object
      required:
      - title
      title: Body_file_anonymous_complaint_api_complaint_anonymous_post
...
    CitizenResolutionRequest:
      properties:
        accept:
          type: boolean
          title: Accept
        note:
          type: string
          title: Note
      type: object
      required:
      - accept
      title: CitizenResolutionRequest
...
    CommissionerComplaintDetailResponse:
      properties:
        complaint:
          $ref: '#/components/schemas/ComplaintSchema'
        updates:
          items:
            $ref: '#/components/schemas/ComplaintUpdateSchema'
          type: array
          title: Updates
      type: object
      required:
      - complaint
      - updates
      title: CommissionerComplaintDetailResponse
...
    ComplaintSchema:
      properties:
        id:
          type: integer
          title: Id
        token:
          type: string
          title: Token
        relatedTokens:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Relatedtokens
        assignedOfficerId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assignedofficerid
        assignedOfficerName:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignedofficername
        departmentId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Departmentid
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        submittedPhotos:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Submittedphotos
        status:
          type: string
          title: Status
        severity:
          type: string
          title: Severity
        resolutionPhotos:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Resolutionphotos
        resolutionNote:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolutionnote
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
        updatedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updatedat
        resolvedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Resolvedat
        closedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Closedat
      type: object
      required:
      - id
      - token
      - title
      - description
      - status
      - severity
      title: ComplaintSchema
    ComplaintUpdateSchema:
      properties:
        id:
          type: integer
          title: Id
        oldStatus:
          anyOf:
          - type: string
          - type: 'null'
          title: Oldstatus
        newStatus:
          type: string
          title: Newstatus
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
      type: object
      required:
      - id
      - newStatus
      title: ComplaintUpdateSchema
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer